% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lc.R
\name{lc_layers}
\alias{lc_layers}
\title{List layers in a Lake County ArcGIS service}
\usage{
lc_layers(folder, service = folder, type = "MapServer")
}
\arguments{
\item{folder}{Folder name (e.g. "Health", "DOT")}

\item{service}{Service name (default: same as folder)}

\item{type}{Service type (default "MapServer")}
}
\value{
tibble: id, name, type
}
\description{
List layers in a Lake County ArcGIS service
}
