% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lc.R
\name{lc_search}
\alias{lc_search}
\title{Search Lake County Hub datasets}
\usage{
lc_search(q = "")
}
\arguments{
\item{q}{Search query}
}
\value{
tibble: id, type, title, description, url
}
\description{
Search Lake County Hub datasets
}
